v2.0.4: Fix Roslyn DLL conflict with Dynamo via AssemblyLoadContext isolation
- Completely removed static references to Microsoft.CodeAnalysis types
- CodeExecutor now loads Roslyn assemblies via custom AssemblyLoadContext (NET8)
- All Roslyn types resolved purely through reflection (no compile-time binding)
- Prevents MissingMethodException when Dynamo or other add-ins load older Roslyn
- Bumped version to 2.0.4 across csproj, package.json, and install.ps1

+++ revit-mcp-plugin/RevitMCPPlugin/Core/CommandExecutor.Cleanup.cs
@@ -396,7 +396,7 @@ namespace RevitMCPPlugin.Core
                 foreach (var vpId in sheet.GetAllViewports())
                 {
                     var vp = doc.GetElement(vpId) as Viewport;
-                    if (vp != null) sheetsWithViewIds.Add((int)vp.ViewId.Value);
+                    if (vp != null) sheetsWithViewIds.Add((int)(long)vp.ViewId.Value);
                 }
             }
